% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examp.R
\name{Bobble}
\alias{Bobble}
\title{Bobble}
\usage{
Bobble(L, n)
}
\arguments{
\item{L}{The matrix with n rows and 1 column for sorting}

\item{n}{The rows of the matrix}
}
\value{
The index of results
}
\description{
Bobble
}
\examples{
\dontrun{
L<-matrix(rnorm(20,1,4),nrow=20)
n<-20
LL<-numeric(n)
Ind<-Bobble(L,n)
Res<-numeric(n)
for ( k in(1:n)){
  ids<-Ind[k]
  Res[k]<-L[ids]
  LL[k]<-L[k,1]
}
print("the original order is:\n")
print(LL)
print("the sorted order is:\n")
print(Res)
}
}
