% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examp.R
\name{fac}
\alias{fac}
\title{compute factorials under recursion}
\usage{
fac(n)
}
\arguments{
\item{n}{for n in n!}
}
\value{
n the value of n!
}
\description{
compute factorials under recursion
}
\examples{
\dontrun{
y<-fac(5)
}
}
