% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examp.R
\name{fEcount}
\alias{fEcount}
\title{the multivariate Epanechnikov (spherical)}
\usage{
fEcount(a1, a2, b1, b2, ha, hb)
}
\arguments{
\item{a1}{with b1 to determine the bar of x}

\item{a2}{with b2 to determine the bar of y}

\item{b1}{with a1 to determine the bar of x}

\item{b2}{with a2 to determine the bar of y}

\item{ha}{the bandwidth of x}

\item{hb}{the bandwidth of y}
}
\value{
the result of  the multivariate Epanechnikov (spherical) kernel
}
\description{
calculate the distribution based on the multivariate Epanechnikov  kernel
}
\examples{
\dontrun{
library(ks)
library(rgl)
xx <- as.matrix(faithful)
kdex <- kde(xx)
data("faithful")
ef<-faithful$eruptions
wf<-faithful$waiting
h1<-0.3
h2<-6
x<-unlist(kdex$eval.points[1])
y<-unlist(kdex$eval.points[2])
n<-length(ef)
n1<-length(x)
n2<-length(y)
z2<-matrix(rep(0,n1*n2),nrow=n1)
for ( i in 1:n1){
  for (j in 1:n2){
    k2<-0
    for (l in 1:n){
      k2<-k2+fEcount(ef[l],wf[l],x[i],y[j],h1,h2)
    }
    z2[i,j]<-k2/n
  }
}
persp(x, y, z2,phi = 30, theta = 40, col = 'red', border = 0,
     xlab='eruptions',ylab='waiting'
      ,main='Results under  Multivariate Epanechnikov (spherical)')
 }
}
