% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Metropolis}
\alias{Metropolis}
\title{A Gibbs sampler using Rcpp}
\usage{
Metropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{the sigma value}

\item{x0}{the start point}

\item{N}{the number of circles}
}
\value{
a matrix of results
}
\description{
A Gibbs  sampler using Rcpp
}
